ADD: Separación de logica en los acchivos principales
